src/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
python-dotenv===1.1.1
pydantic-settings==2.11.0
aiofiles==24.1.0
